+++ pkg/reconciler/gc/reconciler_test.go
@@ -18,7 +18,6 @@ package gc
 
 import (
 	"context"
-	"fmt"
 	"testing"
 	"time"
 
@@ -27,7 +26,6 @@ import (
 	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apimachinery/pkg/runtime/schema"
 	clientgotesting "k8s.io/client-go/testing"
-	duckv1 "knative.dev/pkg/apis/duck/v1"
 	"knative.dev/pkg/configmap"
 	"knative.dev/pkg/controller"
 	"knative.dev/pkg/logging"
@@ -39,7 +37,6 @@ import (
 	gcconfig "knative.dev/serving/pkg/gc"
 	"knative.dev/serving/pkg/reconciler/configuration/resources"
 	"knative.dev/serving/pkg/reconciler/gc/config"
-	gcv1 "knative.dev/serving/pkg/reconciler/gc/v1"
 
 	_ "knative.dev/serving/pkg/client/injection/informers/serving/v1/configuration/fake"
 	_ "knative.dev/serving/pkg/client/injection/informers/serving/v1/revision/fake"
@@ -201,127 +198,6 @@ func TestGCReconcile(t *testing.T) {
 					},
 				}})
 	}))
-}
-
-func TestIsRevisionStale(t *testing.T) {
-	curTime := time.Now()
-	staleTime := curTime.Add(-10 * time.Minute)
-
-	tests := []struct {
-		name      string
-		rev       *v1.Revision
-		latestRev string
-		want      bool
-	}{{
-		name: "fresh revision that was never pinned",
-		rev: &v1.Revision{
-			ObjectMeta: metav1.ObjectMeta{
-				Name:              "myrev",
-				CreationTimestamp: metav1.NewTime(curTime),
-			},
-		},
-		want: false,
-	}, {
-		name: "stale revision that was never pinned w/ Ready status",
-		rev: &v1.Revision{
-			ObjectMeta: metav1.ObjectMeta{
-				Name:              "myrev",
-				CreationTimestamp: metav1.NewTime(staleTime),
-			},
-			Status: v1.RevisionStatus{
-				Status: duckv1.Status{
-					Conditions: duckv1.Conditions{{
-						Type:   v1.RevisionConditionReady,
-						Status: "True",
-					}},
-				},
-			},
-		},
-		want: false,
-	}, {
-		name: "stale revision that was never pinned w/o Ready status",
-		rev: &v1.Revision{
-			ObjectMeta: metav1.ObjectMeta{
-				Name:              "myrev",
-				CreationTimestamp: metav1.NewTime(staleTime),
-			},
-			Status: v1.RevisionStatus{
-				Status: duckv1.Status{
-					Conditions: duckv1.Conditions{{
-						Type:   v1.RevisionConditionReady,
-						Status: "Unknown",
-					}},
-				},
-			},
-		},
-		want: true,
-	}, {
-		name: "stale revision that was previously pinned",
-		rev: &v1.Revision{
-			ObjectMeta: metav1.ObjectMeta{
-				Name:              "myrev",
-				CreationTimestamp: metav1.NewTime(staleTime),
-				Annotations: map[string]string{
-					"serving.knative.dev/lastPinned": fmt.Sprintf("%d", staleTime.Unix()),
-				},
-			},
-		},
-		want: true,
-	}, {
-		name: "fresh revision that was previously pinned",
-		rev: &v1.Revision{
-			ObjectMeta: metav1.ObjectMeta{
-				Name:              "myrev",
-				CreationTimestamp: metav1.NewTime(staleTime),
-				Annotations: map[string]string{
-					"serving.knative.dev/lastPinned": fmt.Sprintf("%d", curTime.Unix()),
-				},
-			},
-		},
-		want: false,
-	}, {
-		name: "stale latest ready revision",
-		rev: &v1.Revision{
-			ObjectMeta: metav1.ObjectMeta{
-				Name:              "myrev",
-				CreationTimestamp: metav1.NewTime(staleTime),
-				Annotations: map[string]string{
-					"serving.knative.dev/lastPinned": fmt.Sprintf("%d", staleTime.Unix()),
-				},
-			},
-		},
-		latestRev: "myrev",
-		want:      false,
-	}}
-
-	cfgStore := testConfigStore{
-		config: &config.Config{
-			RevisionGC: &gcconfig.Config{
-				StaleRevisionCreateDelay:        5 * time.Minute,
-				StaleRevisionTimeout:            5 * time.Minute,
-				StaleRevisionMinimumGenerations: 2,
-			},
-		},
-	}
-	ctx := cfgStore.ToContext(context.Background())
-
-	for _, test := range tests {
-		t.Run(test.name, func(t *testing.T) {
-			cfg := &v1.Configuration{
-				Status: v1.ConfigurationStatus{
-					ConfigurationStatusFields: v1.ConfigurationStatusFields{
-						LatestReadyRevisionName: test.latestRev,
-					},
-				},
-			}
-
-			got := gcv1.IsRevisionStale(ctx, test.rev, cfg)
-
-			if got != test.want {
-				t.Errorf("IsRevisionStale want %v got %v", test.want, got)
-			}
-		})
-	}
 }
 
 func cfg(name, namespace string, generation int64, co ...ConfigOption) *v1.Configuration {
