+++ pkg/reconciler/gc/v1/gc.go
@@ -61,7 +61,7 @@ func Collect(
 	})
 
 	for _, rev := range revs[gcSkipOffset:] {
-		if IsRevisionStale(ctx, rev, config) {
+		if isRevisionStale(ctx, rev, config) {
 			err := client.ServingV1().Revisions(rev.Namespace).Delete(rev.Name, &metav1.DeleteOptions{})
 			if err != nil {
 				logger.With(zap.Error(err)).Errorf("Failed to delete stale revision %q", rev.Name)
@@ -72,7 +72,7 @@ func Collect(
 	return nil
 }
 
-func IsRevisionStale(ctx context.Context, rev *v1.Revision, config *v1.Configuration) bool {
+func isRevisionStale(ctx context.Context, rev *v1.Revision, config *v1.Configuration) bool {
 	if config.Status.LatestReadyRevisionName == rev.Name {
 		return false
 	}
